Introduce new `iced_highlighter` subcrate
