+++ examples/editor/Cargo.toml
@@ -7,7 +7,7 @@ publish = false
 
 [dependencies]
 iced.workspace = true
-iced.features = ["advanced", "tokio", "debug"]
+iced.features = ["highlighter", "tokio", "debug"]
 
 tokio.workspace = true
 tokio.features = ["fs"]
